+++ src/torch_pointcloud/models/pvcnn.py
@@ -471,7 +471,7 @@ class PVCNNSegmentation(SegmentationModel):
     "pvcnn.s3dis-area5.mit-han-lab",
     task="segmentation",
     weights=WeightsDict(
-        url="hf://torch-pointcloud/pvcnn.s3dis-area5.mit-han-lab/resolve/main/model.safetensors",
+        url="hf://torch-pointcloud/pvcnn.s3dis-area5.mit-han-lab/resolve/a098f49f11480c59e19b14cebb8f40df11fdccdc/model.safetensors",
         dataset="s3dis-area5",
         metrics={"mIoU": 57.51, "OA": 86.58},
         classes=S3DIS_CLASSES,